@@ -3,54 +3,47 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@drizzle/react-components": "^1.5.3",
-    "@drizzle/react-plugin": "^1.5.3",
-    "@drizzle/store": "^1.5.3",
-    "@ethereumjs/common": "^2.1.0",
+    "@ethereumjs/tx": "^3.0.2",
     "@material-ui/core": "^4.11.3",
     "@material-ui/icons": "^4.11.2",
-    "@transmute/element-lib": "^0.8.4",
-    "@transmute/universal-wallet": "^0.4.1",
-    "@veramo/key-manager": "^1.1.0",
+    "@testing-library/jest-dom": "^5.11.4",
+    "@testing-library/react": "^11.1.0",
+    "@testing-library/user-event": "^12.1.10",
+    "@textile/hub": "^6.1.2",
+    "@truffle/contract": "^4.3.10",
     "axios": "^0.21.1",
+    "bcryptjs": "^2.4.3",
+    "bip39": "^3.0.3",
     "bl": "^4.1.0",
-    "bn.js": "^5.1.3",
-    "buffer": "^6.0.3",
-    "crypto": "^1.0.1",
+    "bn.js": "^5.2.0",
     "crypto-js": "^4.0.0",
-    "did-jwt": "^4.8.1",
-    "did-resolver": "^2.1.2",
-    "drizzle-react": "^1.3.0",
+    "did-jwt": "^5.0.1",
+    "did-resolver": "^3.0.2",
+    "eccrypto": "^1.1.5",
+    "eccrypto-js": "^5.4.0",
     "elliptic": "^6.5.4",
-    "ethereum-public-key-to-address": "0.0.5",
+    "eth-sig-util": "^3.0.1",
+    "ethereum-public-key-to-address": "^0.0.5",
+    "ethereumjs-abi": "^0.6.8",
     "ethereumjs-common": "^1.5.2",
-    "ethers": "^5.0.31",
-    "ethjs-abi": "^0.2.1",
-    "ethjs-contract": "^0.2.3",
-    "ethjs-query": "^0.3.8",
+    "ethereumjs-tx": "^2.1.2",
+    "ethereumjs-util": "^7.0.9",
+    "ethereumjs-wallet": "^1.0.1",
+    "ethers": "^5.0.32",
     "ethr-did": "^1.3.0",
-    "ethr-did-registry": "0.0.3",
+    "ethr-did-registry": "^0.0.3",
     "ethr-did-resolver": "^3.0.3",
-    "ipfs-http-client": "^49.0.1",
-    "ipns": "^0.8.2",
-    "jose": "^3.6.0",
-    "js-sha3": "^0.8.0",
-    "node-couchdb": "^1.3.0",
-    "npm": "^7.5.2",
-    "react": "16.11.0",
-    "react-components": "^0.5.1",
-    "react-dom": "16.11.0",
+    "ipfs-http-client": "^49.0.4",
+    "react": "^17.0.1",
+    "react-dom": "^17.0.1",
     "react-material-ui-form-validator": "^2.1.4",
     "react-router-dom": "^5.2.0",
-    "react-scripts": "3.2.0",
-    "recoil": "^0.1.2",
+    "react-scripts": "4.0.3",
+    "recoil": "^0.1.3",
     "secrets.js": "^0.1.8",
-    "secrets.js-grempe": "^2.0.0",
-    "shamirs-secret-sharing": "^1.0.1",
     "slip39": "^0.1.7",
-    "tabler-react": "^2.0.0-alpha.1",
-    "typescript": "^4.1.3",
-    "web3": "^1.3.4"
+    "web-vitals": "^1.0.1",
+    "web3": "^1.2.4"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -59,7 +52,10 @@
     "eject": "react-scripts eject"
   },
   "eslintConfig": {
-    "extends": "react-app"
+    "extends": [
+      "react-app",
+      "react-app/jest"
+    ]
   },
   "browserslist": {
     "production": [
@@ -72,9 +68,5 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
-  },
-  "devDependencies": {
-    "@babel/plugin-proposal-nullish-coalescing-operator": "^7.13.0",
-    "@types/node": "^14.14.25"
   }
 }
