@@ -1,4 +1,4 @@
-import React, {useEffect} from 'react';
+import React, {useEffect, useState} from 'react';
 import {Switch, Route} from "react-router-dom";
 import ConfirmBIP39Phrase from './pages/sign-up-flow/confirm-bip39-phrase/Confirm-Bip39-Phrase';
 import CreateBIP39 from './pages/sign-up-flow/create-bip39-phrase/CreateBIP39';
@@ -19,6 +19,13 @@ import VerifyPage from './pages/verify-page/VerifyPage';
 // import {drizzleReactHooks} from "@drizzle/react-plugin";
 import CreateDataVault from './pages/vault/create-data-vault/CreateDataVault';
 import VaultRouter from './pages/vault/vault-shell/VaultRouter';
+import ViewNFTs from './pages/verify-page/view_NFTs';
+
+import LandingPage from "./pages/welcome/Welcome";
+import HomeWrapper from './pages/home/HomeWrapper';
+
+import Web3Context from "./web3Context";
+import ImportVault from './pages/import-vault/ImportVault';
 
 
 // const drizzle = new Drizzle(drizzleOptions);
@@ -30,17 +37,25 @@ import VaultRouter from './pages/vault/vault-shell/VaultRouter';
  * Acts as a general switch board for the rest of the application
  */
 function App(){
-    
-    // useEffect(() => {
-    //     const web3 = getWeb3();
-    //     // provide web3 for all required services
-    //     web3Service.initWeb3(web3);
-    // }, []);
+    const [web3, setWeb3] = useState(undefined);
+    useEffect(() => {
+        const runEffect = async () => {
+            const w3 = await getWeb3();
+            setWeb3(w3);
+        }
+        runEffect();
+    }, []);
 
     return (
         <div className="App">
+            <Web3Context.Provider value={{web3}}>
             {/* <DrizzleProvider drizzle={drizzle}> */}
                 <Switch>
+                    {/* Welcome Page */}
+                    <Route exact path="/" render={() => <LandingPage/>}/>
+
+                    <Route path="/home" render={() => <HomeWrapper/>}/>
+
                     {/* For current test purpsoes */}
                     <Route path="/setup" render={() => <OnboardingShell/>}/>
 
@@ -50,8 +65,14 @@ function App(){
                     {/* Vault stuff */}
                     <Route path="/vault" render={() => <VaultRouter/>}/>
 
+                    {/* Import Vault */}
+                    <Route path="/import" render={() => <ImportVault/>} />
+
                     {/* For when a user wishes to verify themselves as a trustee */}
                     <Route path="/verify" render={() => <VerifyPage/>}/>
+
+                    {/* For a user who wishes to verify to view their NFTs */}
+                    <Route path="/viewnfts" render={() => <ViewNFTs/>}/>
 
                     {/* For testing creating the vault */}
                     <Route path="/create-vault" render={() => <CreateDataVault/>}/>
@@ -63,6 +84,7 @@ function App(){
                     {/* <Route exact path="/" render={() => <HandlePathing />} /> */}
                 </Switch>
             {/* </DrizzleProvider> */}
+            </Web3Context.Provider>
         </div>
     );
 }
